requirements.txt
catboost
entrypoints
flake8
numpy
pandas
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
matplotlib
matplotlib-inline
seaborn
scikit-learn
hydra-core
pytest
requests
httpx
fastapi
"uvicorn[standard]"